Remove fonts and nodejs
While the runtime by default ships no CJK fonts, such fonts are always
exposed from the host by flatpak. There is no point in installing a
duplicate set of fonts in the flatpak

nodejs 0.12 went out of support a long time ago and it fails to build
with the current runtime.

The only use for it was writing extensions for Notepadqq and the last
upstream position is that they are not too keen on supporting it anymore

https://github.com/notepadqq/notepadqq/issues/686

Remove unused patches and comments